+++ src/main/java/com/example/jpa/domain/point/repository/PointRepository.java
@@ -7,5 +7,5 @@ import java.util.List;
 
 public interface PointRepository extends JpaRepository<Point, Integer> {
     // 특정 사용자의 포인트 내역을 최신순으로 조회
-    List<Point> findByUserIdOrderByCreatedAtDesc(Integer userId);
+    List<Point> findByUserIdOrderByCreatedAtDesc(Long userId);
 }
